Move some things to utilities
